+++ all-is-cubes-content/src/city/exhibits/move_modifier.rs
@@ -16,7 +16,7 @@ fn MOVED_BLOCKS(_: Context<'_>) {
             let i = x + z * 8;
             let distance = (i * 16).try_into().unwrap();
             let block = &blocks[i as usize];
-            let [move_out, move_in] = Move::new(Face6::PY, distance, 0).to_paired();
+            let [move_out, move_in] = Move::new(Face6::PY, distance, 0).into_paired();
             // TODO: Move should be able to spawn a "tail" on its own when animated?
             space.set(
                 [x * 2, 0, (1 - z) * 2],
@@ -28,7 +28,7 @@ fn MOVED_BLOCKS(_: Context<'_>) {
             )?;
 
             // Horizontal
-            let [move_out, move_in] = Move::new(Face6::PZ, distance, 0).to_paired();
+            let [move_out, move_in] = Move::new(Face6::PZ, distance, 0).into_paired();
             space.set([i, 0, -2], block.clone().with_modifier(move_out))?;
             space.set([i, 0, -1], block.clone().with_modifier(move_in))?;
         }
@@ -50,7 +50,7 @@ fn PROJECTILE(ctx: Context<'_>) {
 
     let projectile = &demo_blocks[DemoBlocks::Projectile];
 
-    let [move_in, _move_out] = Move::new(Face6::NY, 256, -32).to_paired();
+    let [move_in, _move_out] = Move::new(Face6::NY, 256, -32).into_paired();
     let projectile_moving_in = projectile.clone().with_modifier(move_in);
     // let projectile_moving_out = projectile.clone().with_modifier(move_out);